+++ CHANGELOG.md
@@ -3,5 +3,8 @@
 * (for developers) Block Registry API!
     * Now developers can register their own whitelisted blocks with the gadget!
 
+## Changed
+* Gadget bauble support is now indicated in a tooltip
+
 ## Fixes
 * Redo operations with ArchitectureCraft blocks don't preserve block data